rename JS module. Set data-bakdrop="static" on form modal...
Making a modal's backdrop static via data-backdrop="static" and
data-keyboard="false" prevent's the modal from closing when clicking
outside the window.

This is a nice feature when filling out a form, b/c it is easy to
inadvertantly click outside of a narrow modal (closing the modal window)
while trying to click into form fields and other form elements.